More moving around of client code. Better exception formatting. More precision on altitude.

+++ ExtLibs/AltitudeAngelWings/Models/FlightDataPosition.cs
@@ -5,7 +5,7 @@ namespace AltitudeAngelWings.Models
         public double Longitude { get; set; }
         public double Latitude { get; set; }
         public double Speed { get; set; }
-        public int Altitude { get; set; }
+        public double Altitude { get; set; }
         public float Course { get; set; }
         public float VerticalSpeed { get; set; }
     }
